+++ testproject/testproject/models.py
@@ -3,8 +3,8 @@ from django.db.models.signals import post_save, post_delete
 
 from mptt.models import TreeManager
 
-from groups_manager.models import Group, Member, GroupType, \
-    group_save, group_delete, member_save, member_delete
+from groups_manager.models import Group, Member, GroupType, GroupMember, \
+    group_save, group_delete, member_save, member_delete, group_member_save, group_member_delete
 
 '''
 Legion main example
@@ -64,6 +64,71 @@ class Newsletter(models.Model):
     class Meta:
         permissions = (('view_newsletter', 'View Newsletter'),
                        ('send_newsletter', 'Send Newsletter'), )
+
+
+'''
+Signals kwargs example with subclass
+'''
+
+
+class ProjectGroup(Group):
+
+    class Meta:
+        permissions = (('view_projectgroup', 'View Project Group'), )
+
+    class GroupsManagerMeta:
+        member_model = 'testproject.ProjectGroupMember'
+        group_member_model = 'testproject.ProjectGroupMember'
+
+
+class ProjectMember(Member):
+
+    class Meta:
+        permissions = (('view_projectmember', 'View Project Member'), )
+
+
+class ProjectGroupMember(GroupMember):
+    pass
+
+
+def project_get_auth_models_sync_func(instance):
+    return True
+
+
+def project_group_save(*args, **kwargs):
+    group_save(*args, get_auth_models_sync_func=project_get_auth_models_sync_func,
+               prefix='PGS_', suffix='_Project', **kwargs)
+
+
+def project_group_delete(*args, **kwargs):
+    group_delete(*args, get_auth_models_sync_func=project_get_auth_models_sync_func, **kwargs)
+
+
+def project_member_save(*args, **kwargs):
+    member_save(*args, get_auth_models_sync_func=project_get_auth_models_sync_func,
+                prefix='PGS_', suffix='_Project', **kwargs)
+
+
+def project_member_delete(*args, **kwargs):
+    member_delete(*args, get_auth_models_sync_func=project_get_auth_models_sync_func, **kwargs)
+
+
+def project_group_member_save(*args, **kwargs):
+    group_member_save(*args, get_auth_models_sync_func=project_get_auth_models_sync_func, **kwargs)
+
+
+def project_group_member_delete(*args, **kwargs):
+    group_member_delete(*args, get_auth_models_sync_func=project_get_auth_models_sync_func, **kwargs)
+
+
+post_save.connect(project_group_save, sender=ProjectGroup)
+post_delete.connect(project_group_delete, sender=ProjectGroup)
+
+post_save.connect(project_member_save, sender=ProjectMember)
+post_delete.connect(project_member_delete, sender=ProjectMember)
+
+post_save.connect(project_group_member_save, sender=ProjectGroupMember)
+post_delete.connect(project_group_member_delete, sender=ProjectGroupMember)
 
 
 '''